+++ Radi-core/src/graphics/font.h
@@ -14,12 +14,12 @@ namespace radi
 		private:
 			ftgl::texture_atlas_t* m_FTAtlas;
 			ftgl::texture_font_t* m_FTFont;
-			unsigned int m_size;
+			float m_size;
 			maths::vec2 m_scale;
 			std::string m_name;
 			std::string m_filename;
 		public:
-			Font(std::string name, std::string filename, int size);
+			Font(std::string name, std::string filename, float size);
 
 			void setScale(float x, float y);
 
@@ -29,7 +29,7 @@ namespace radi
 			inline const maths::vec2 getScale() const { return m_scale; }
 			inline const std::string getName() const { return m_name; }
 			inline const std::string getFilename() const { return m_filename; }
-			inline const unsigned int getSize() const { return m_size; }
+			inline const float getSize() const { return m_size; }
 		};
 	}
 }
