+++ Radi-core/src/maths/maths_func.h
@@ -9,12 +9,12 @@ namespace radi
 	{
 		inline float toRadians(float degrees)
 		{
-			return degrees * (M_PI / 180.0f);
+			return (float)(degrees * (M_PI / 180.0f));
 		}
 
 		inline float toDegrees(float radians)
 		{
-			return radians * (180.0f / M_PI);
+			return (float)(radians * (180.0f / M_PI));
 		}
 
 		inline int sign(float value)
